+++ pkg/validation/app_test.go
@@ -571,6 +571,39 @@ func Test_ValidateApp(t *testing.T) {
 			},
 			expectedErr: "validation error: label `app-operator.giantswarm.io/version` has invalid value `1.0.0`",
 		},
+		{
+			name: "case 18: nginx user values configmap name is not restricted",
+			obj: v1alpha1.App{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "kiam",
+					Namespace: "eggs2",
+					Labels: map[string]string{
+						label.AppOperatorVersion: "0.0.0",
+					},
+				},
+				Spec: v1alpha1.AppSpec{
+					Catalog:   "default",
+					Name:      "nginx-ingress-controller-app",
+					Namespace: "kube-system",
+					KubeConfig: v1alpha1.AppSpecKubeConfig{
+						InCluster: true,
+					},
+					UserConfig: v1alpha1.AppSpecUserConfig{
+						ConfigMap: v1alpha1.AppSpecUserConfigConfigMap{
+							Name:      "nginx-ingress-user-values",
+							Namespace: "eggs2",
+						},
+					},
+					Version: "1.4.0",
+				},
+			},
+			catalogs: []*v1alpha1.Catalog{
+				newTestCatalog("default", "default"),
+			},
+			configMaps: []*corev1.ConfigMap{
+				newTestConfigMap("nginx-ingress-user-values", "eggs2"),
+			},
+		},
 	}
 
 	for _, tc := range tests {